Added table rows, columns and striped
PRNJ-2831
