+++ Storybook/Stories/Table/index.js
@@ -1,0 +1,6 @@
+import {storiesOf} from '@storybook/vue';
+import UseCases from './UseCases.story.vue';
+
+export default () => {
+  storiesOf('Table', module).add('Use Cases', () => UseCases);
+};
