+++ Storybook/Stories/index.js
@@ -1,7 +1,9 @@
 import storyOfIcon from 'Stories/Icon';
 import storyOfTag from 'Stories/Tag';
+import storyOfTable from 'Stories/Table';
 
 export default () => {
   storyOfIcon();
   storyOfTag();
+  storyOfTable();
 };
