+++ src/styles/components/index.less
@@ -6,6 +6,7 @@
 @import 'select';
 @import 'select-dropdown';
 @import 'tag';
+@import 'table';
 
 /* @import 'affix';
 @import 'back-top';
@@ -33,7 +34,6 @@
 @import 'slider';
 @import 'cascader';
 @import 'transfer';
-@import 'table';
 @import 'dropdown';
 @import 'tabs';
 @import 'menu';
