+++ src/styles/custom.less
@@ -28,6 +28,26 @@
 @font-weight-bolder: 800;
 @line-height-normal: 1.5rem;
 
+// TextSizes
+@textSizes-11: 0.707rem;
+@textSizes-14: 0.871rem;
+@textSizes-16: 1rem;
+@textSizes-22: 1.414rem;
+@textSizes-32: 2rem;
+@textSizes-45: 2.828rem;
+@textSizes-64: 4rem;
+
+// FontWeights
+@fontWeights-hairline: 100;
+@fontWeights-thin: 200;
+@fontWeights-light: 300;
+@fontWeights-normal: 400;
+@fontWeights-medium: 500;
+@fontWeights-semibold: 600;
+@fontWeights-bold: 700;
+@fontWeights-extrabold: 800;
+@fontWeights-black: 900;
+
 // Borders
 @border-color-base: @color-s500;
 
@@ -35,9 +55,24 @@
 @radius-normal: 0.25rem;
 @radius-big: 0.5rem;
 @radius-rounded: 0.75rem;
+@rounded-default: .25rem;
+
+// Sizes
+@sizes-4: 0.25rem;
+@sizes-8: 0.5rem;
+@sizes-12: 0.75rem;
+@sizes-16: 1rem;
+@sizes-24: 1.5rem;
+@sizes-32: 2rem;
+@sizes-40: 2.5rem;
+@sizes-48: 3rem;
+@sizes-56: 3.5rem;
+@sizes-64: 4rem;
+@sizes-auto: auto;
 
 // Shadows
 @shadow-small: 0 2px 12px fade(@color-dark100, 6%), 0 1px 2px fade(@color-dark100, 6%);
+@shadows-default: 0 2px 12px rgba(36, 28, 25, 0.06), 0 1px 2px rgba(36, 28, 25, 0.06);
 
 // Animation
 @animation-time: 0.3s;
